fix: refactor taggedlen/rawsmallvec and visibility
